requirements.txt
# Core Dependencies
flask==3.0.0
requests==2.31.0
pillow==10.1.0
python-dotenv==1.0.0

# LLM Integration
openai==1.3.0

# Optional: Local LLM support
# transformers==4.35.0
# torch==2.1.0

# Data handling
numpy==1.26.2

# Web server
gunicorn==21.2.0
